requirements.txt
asgiref==3.2.7
Django==3.0.6
django-filter==2.2.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.6
pytz==2020.1
selenium==3.141.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
